Move some functions to their own files
