+++ src/commonMain/kotlin/GetMarketBoardSaleHistory.kt
@@ -1,0 +1,87 @@
+package cloud.drakon.ktuniversalis
+
+import cloud.drakon.ktuniversalis.entities.History
+import cloud.drakon.ktuniversalis.entities.Multi
+import io.ktor.client.call.body
+import io.ktor.client.request.get
+
+/**
+ * Returns the history data for the requested set of item IDs and world or data center.
+ * @param worldDcRegion The world or data center to retrieve data for. This may be an ID or a name. Regions should be specified as Japan, Europe, North-America, Oceania, China, or 中国.
+ * @param itemIds The set of item IDs to retrieve data for.
+ * @param entriesToReturn The number of entries to return. By default, this is set to `1800`, but may be set to a maximum of `999999`.
+ * @param statsWithin The amount of time before now to calculate stats over, in milliseconds. By default, this is `7` days.
+ * @param entriesWithin The amount of time before now to take entries within, in seconds. Negative values will be ignored.
+ * @throws UniversalisException The Universalis API returned an unexpected return code.
+ */
+private suspend fun getMarketBoardSaleHistorySet(
+    worldDcRegion: String,
+    itemIds: Set<Int>,
+    entriesToReturn: Int? = null,
+    statsWithin: Int? = null,
+    entriesWithin: Int? = null,
+): Any = ktorClient.get(
+    "history/$worldDcRegion/" + itemIds.joinToString(",")
+) {
+    url {
+        if (entriesToReturn != null) parameters.append(
+            "entriesToReturn", entriesToReturn.toString()
+        )
+
+        if (statsWithin != null) parameters.append(
+            "statsWithin", statsWithin.toString()
+        )
+
+        if (entriesWithin != null) parameters.append(
+            "entriesWithin", entriesWithin.toString()
+        )
+    }
+}.let {
+    when (it.status.value) {
+        200  -> if (itemIds.size == 1) {
+            it.body() as History
+        } else {
+            it.body() as Multi<History>
+        }
+
+        else -> throw throwUniversalisException(it)
+    }
+}
+
+/**
+ * Returns the history data for the requested item ID and world or data center.
+ * @param worldDcRegion The world or data center to retrieve data for. This may be an ID or a name. Regions should be specified as Japan, Europe, North-America, Oceania, China, or 中国.
+ * @param itemId The item ID to retrieve data for.
+ * @param entriesToReturn The number of entries to return. By default, this is set to `1800`, but may be set to a maximum of `999999`.
+ * @param statsWithin The amount of time before now to calculate stats over, in milliseconds. By default, this is `7` days.
+ * @param entriesWithin The amount of time before now to take entries within, in seconds. Negative values will be ignored.
+ * @throws UniversalisException The Universalis API returned an unexpected return code.
+ */
+suspend fun getMarketBoardSaleHistory(
+    worldDcRegion: String,
+    itemId: Int,
+    entriesToReturn: Int? = null,
+    statsWithin: Int? = null,
+    entriesWithin: Int? = null,
+): History = getMarketBoardSaleHistorySet(
+    worldDcRegion, setOf(itemId), entriesToReturn, statsWithin, entriesWithin
+) as History
+
+/**
+ * Returns the history data for the requested set of item IDs and world or data center.
+ * @param worldDcRegion The world or data center to retrieve data for. This may be an ID or a name. Regions should be specified as Japan, Europe, North-America, Oceania, China, or 中国.
+ * @param itemIds The set of item IDs to retrieve data for.
+ * @param entriesToReturn The number of entries to return. By default, this is set to `1800`, but may be set to a maximum of `999999`.
+ * @param statsWithin The amount of time before now to calculate stats over, in milliseconds. By default, this is `7` days.
+ * @param entriesWithin The amount of time before now to take entries within, in seconds. Negative values will be ignored.
+ * @throws UniversalisException The Universalis API returned an unexpected return code.
+ */
+suspend fun getMarketBoardSaleHistory(
+    worldDcRegion: String,
+    itemIds: Set<Int>,
+    entriesToReturn: Int? = null,
+    statsWithin: Int? = null,
+    entriesWithin: Int? = null,
+): Multi<History> = getMarketBoardSaleHistorySet(
+    worldDcRegion, itemIds, entriesToReturn, statsWithin, entriesWithin
+) as Multi<History>
